daca/requirements.txt
asgiref==3.3.1
DateTime==4.3
dj-rest-auth==2.1.2
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
importlib-metadata==3.1.1
Markdown==3.3.3
pytz==2020.4
sqlparse==0.4.1
zipp==3.4.0
zope.interface==5.2.0
